+++ framework/server/src/main/java/module-info.java
@@ -28,7 +28,7 @@ module is.codion.framework.server {
 	requires org.slf4j;
 	requires is.codion.tools.jul.classpath;
 	requires is.codion.framework.db.local;
-	requires transitive is.codion.framework.db.rmi;
+	requires is.codion.framework.db.rmi;
 
 	exports is.codion.framework.server;
 }